requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
Django==2.0.5
django-configurations==2.0
django-debug-toolbar==1.9.1
djangorestframework==3.8.2
Fabric3==1.14.post1
Faker==0.8.15
idna==2.6
invoke==1.0.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
paramiko==2.4.1
Pillow==5.1.0
psycopg2-binary==2.7.4
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
